requirements.txt
opencv-python
onnxruntime 
mysql-connector-python